fixup??? Add reftable library
Do close the file descriptors to obsolete files before trying to delete
them. This is actually required on Windows.

Note: this patch is probably a bit heavy-handed, releasing more than
necessary (which will then have to be re-read). But it seems to be
enough to let t0031 pass on Windows, so it is at least a clue as to what
the fix will look like eventually.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>